CDMS-366: code review changes
