+++ package.json
@@ -57,12 +57,10 @@
     "ioredis": "^5.4.1",
     "joi": "^17.4.2",
     "kitsu-core": "^10.2.0",
-    "lodash": "^4.17.21",
     "nunjucks": "^3.2.4",
     "pino": "^9.5.0",
     "pino-pretty": "^13.0.0",
-    "undici": "^6.20.1",
-    "uuid": "^11.1.0"
+    "undici": "^6.20.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.25.9",
