--- src/plugins/auth/drop-user-session.js
@@ -1,5 +1,0 @@
-function dropUserSession () {
-  return this.server.app.cache.drop(this.state.userSession.sessionId)
-}
-
-export { dropUserSession }
