--- src/plugins/auth/get-user-session.js
@@ -1,7 +1,0 @@
-async function getUserSession () {
-  return this.state?.userSession?.sessionId
-    ? this.server.app.cache.get(this.state.userSession.sessionId)
-    : {}
-}
-
-export { getUserSession }
