+++ src/plugins/auth/session-cookie.js
@@ -1,55 +1,9 @@
 import authCookie from '@hapi/cookie'
-import { isPast, parseISO, subMinutes } from 'date-fns'
 
 import { config } from '../../config/config.js'
-import { refreshAccessToken } from './refesh-token.js'
-import { removeUserSession, updateUserSession } from './user-session.js'
+import { validateUserSession } from '../../auth/user-session.js'
 
 const sessionConfig = config.get('session')
-
-const validateSession = async (server, request, session) => {
-  const authedUser = await request.getUserSession()
-
-  if (!authedUser) {
-    return { isValid: false }
-  }
-
-  const tokenHasExpired = isPast(
-    subMinutes(parseISO(authedUser.expiresAt), 1)
-  )
-
-  if (tokenHasExpired) {
-    const response = await refreshAccessToken(request)
-
-    if (!response.ok) {
-      removeUserSession(request)
-
-      return { isValid: false }
-    }
-
-    const refreshAccessTokenJson = await response.json()
-    const updatedSession = await updateUserSession(
-      request,
-      refreshAccessTokenJson
-    )
-
-    return {
-      isValid: true,
-      credentials: updatedSession
-    }
-  }
-
-  const userSession = await server.app.cache.get(session.sessionId)
-
-  if (userSession) {
-    return {
-      isValid: true,
-      credentials: userSession
-    }
-  }
-
-  return { isValid: false }
-}
 
 const sessionCookie = {
   plugin: {
@@ -67,7 +21,7 @@ const sessionCookie = {
         },
         keepAlive: true,
         validate: async (request, session) => {
-          return validateSession(server, request, session)
+          return validateUserSession(server, request, session)
         }
       })
 
@@ -77,6 +31,5 @@ const sessionCookie = {
 }
 
 export {
-  validateSession,
   sessionCookie
 }
