+++ src/plugins/template-renderer/context.js
@@ -4,6 +4,7 @@ import { readFileSync } from 'node:fs'
 import { config } from '../../config/config.js'
 import { createLogger } from '../../utils/logger.js'
 import { getNavigationItems } from '../../config/navigation-items.js'
+import { getUserSession } from '../../auth/user-session.js'
 
 const logger = createLogger()
 const assetPath = config.get('assetPath')
@@ -27,7 +28,7 @@ export async function context (request) {
     }
   }
 
-  const authedUser = await request.getUserSession()
+  const authedUser = await getUserSession(request)
 
   return {
     authedUser,
