+++ src/routes/index.js
@@ -2,8 +2,8 @@ import { health } from './health.js'
 import { staticAssetRoutes } from './static-assets.js'
 import { home } from './home.js'
 import { signinOidc } from './signin-oidc.js'
-import { login } from './login.js'
-import { logout } from './logout.js'
+import { signIn } from './sign-in.js'
+import { signOut } from './sign-out.js'
 import { search } from './search.js'
 import { searchResult } from './search-result.js'
 
@@ -15,8 +15,8 @@ const defaultRoutes = [
 
 const appSpecificRoutes = [
   signinOidc,
-  login,
-  logout,
+  signIn,
+  signOut,
   ...search,
   searchResult
 ]
