--- src/routes/prerequisites/provide-authed-user.js
@@ -1,6 +1,0 @@
-const provideAuthedUser = {
-  method: async (request) => request.getUserSession(),
-  assign: 'authedUser'
-}
-
-export { provideAuthedUser }
