+++ src/routes/route-constants.js
@@ -1,8 +1,8 @@
 const paths = {
   LANDING: '/',
-  AUTH: '/signin-oidc',
-  LOGIN: '/login',
-  LOGOUT: '/logout',
+  AUTH_DEFRA_ID_CALLBACK: '/signin-oidc',
+  SIGN_IN: '/sign-in',
+  SIGN_OUT: '/sign-out',
   SEARCH: '/search',
   SEARCH_RESULT: '/search-result'
 }
