+++ src/server.js
@@ -4,8 +4,6 @@ import hapi from '@hapi/hapi'
 import { config } from './config/config.js'
 import plugins from './plugins/index.js'
 import { getCacheEngine } from './utils/caching/cache-engine.js'
-import { getUserSession } from './plugins/auth/get-user-session.js'
-import { dropUserSession } from './plugins/auth/drop-user-session.js'
 
 export async function createServer () {
   const server = hapi.server({
@@ -54,9 +52,6 @@ export async function createServer () {
     expiresIn: config.get('session.cache.ttl'),
     segment: 'session'
   })
-
-  server.decorate('request', 'getUserSession', getUserSession)
-  server.decorate('request', 'dropUserSession', dropUserSession)
 
   await server.register(plugins)
