+++ test/unit/auth/refresh-token.test.js
@@ -1,8 +1,9 @@
-import { refreshAccessToken } from '../../../../src/plugins/auth/refesh-token.js'
-import { createAuthedUser } from '../../utils/session-helper.js'
-import { createLogger } from '../../../../src/utils/logger.js'
+import { refreshAccessToken } from '../../../src/auth/refesh-token.js'
+import { createAuthedUser } from '../utils/session-helper.js'
+import { createLogger } from '../../../src/utils/logger.js'
 import Querystring from 'querystring'
-import { config } from '../../../../src/config/config.js'
+import { config } from '../../../src/config/config.js'
+import { getUserSession } from '../../../src/auth/user-session.js'
 
 const mockPost = jest.fn()
 const mockLoggerInfo = jest.fn()
@@ -13,10 +14,14 @@ jest.mock('@hapi/wreck', () => ({
   post: (...args) => mockPost(...args)
 }))
 
-jest.mock('../../../../src/utils/logger.js', () => ({
+jest.mock('../../../src/utils/logger.js', () => ({
   createLogger: () => ({
     info: (...args) => mockLoggerInfo(...args)
   })
+}))
+
+jest.mock('../../../src/auth/user-session.js', () => ({
+  getUserSession: jest.fn()
 }))
 
 describe('#refreshToken', () => {
@@ -33,10 +38,9 @@ describe('#refreshToken', () => {
       const clientId = authConfig.defraId.clientId
       const clientSecret = authConfig.defraId.clientSecret
 
+      getUserSession.mockReturnValue(authedUser)
+
       const request = {
-        getUserSession: () => {
-          return authedUser
-        },
         logger: logger
       }
 
@@ -45,7 +49,7 @@ describe('#refreshToken', () => {
         client_secret: clientSecret,
         grant_type: 'refresh_token',
         refresh_token: authedUser.refreshToken,
-        scope: `${clientId} openid offline_access`
+        scope: `${clientId} openid`
       })
 
       await refreshAccessToken(request)
