+++ test/unit/routes/signin-oidc.test.js
@@ -1,12 +1,9 @@
-import { v4 as uuidv4 } from 'uuid'
 import { startServer } from '../../../src/utils/start-server.js'
 import { createAuthedUser } from '../utils/session-helper.js'
 import { paths } from '../../../src/routes/route-constants.js'
 import { constants as httpConstants } from 'http2'
 
-jest.mock('uuid', () => ({
-  v4: jest.fn()
-}))
+crypto.randomUUID = () => 'a-test-session-id'
 
 describe('#signinOidc', () => {
   describe('When accessed following successful signin', () => {
@@ -16,8 +13,6 @@ describe('#signinOidc', () => {
       jest.clearAllMocks()
       server = await startServer()
       userSession = createAuthedUser()
-
-      uuidv4.mockReturnValue('a-test-session-id')
     })
 
     afterEach(async () => {
@@ -27,7 +22,7 @@ describe('#signinOidc', () => {
     test('Should redirect to search page', async () => {
       const { statusCode, headers, request } = await server.inject({
         method: 'GET',
-        url: paths.AUTH,
+        url: paths.AUTH_DEFRA_ID_CALLBACK,
         auth: {
           strategy: 'defra-id',
           credentials: userSession
